casts for less warnings
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@85 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ view/special/MegaMolLogo.cpp
@@ -101,7 +101,7 @@ void megamol::core::view::special::MegaMolLogo::Create(void) {
         }
     }
 
-    this->lineIdxCnt = lines.Count() * 2;
+    this->lineIdxCnt = static_cast<unsigned int>(lines.Count() * 2);
     this->lineIdxs = new unsigned int[this->lineIdxCnt];
     for (unsigned int i = 0; i < lines.Count(); i++) {
         this->lineIdxs[i * 2] = lines[i].First();
